requirements.txt
discord.py
dnspython
GitPython
pymongo
lavalink
aiohttp
websockets
aiohttp
pyttanko
redis